started work on static content and theming the site.
Currently no UI has been done, the theming is in the sense of ensuring
we do not need to manually change the header / footer for every view.
Work has also started on some of the static content as well as a minor
change to the config of the site which changes subclass prefixes to MH
over MY. The MH standing for mozhunt.

Also added are a draft privacy policy and draft terms of service… both
will need work and more information needs to be added to the TOS such
as the rules of the game.